remove combo + cleanup
