+++ packages/core/src/index.ts
@@ -2,12 +2,10 @@ require("./polyfills");
 
 import { BaseChart } from "./base-chart";
 import { BaseAxisChart } from "./base-axis-chart";
-
 import { PieChart } from "./pie-chart";
 import { DonutChart, DonutCenter } from "./donut-chart";
 import { BarChart } from "./bar-chart";
 import { LineChart } from "./line-chart";
-import { ComboChart } from "./combo-chart";
 import { ScatterChart } from "./scatter-chart";
 
 import * as colorPalettes from "./services/colorPalettes";
@@ -24,6 +22,5 @@ export {
 	DonutCenter,
 	BarChart,
 	LineChart,
-	ComboChart,
 	ScatterChart
 };
