+++ packages/core/src/scatter-chart.ts
@@ -8,6 +8,7 @@ import { Tools } from "./tools";
 
 export class ScatterChart extends BaseAxisChart {
 	options: ScatterChartOptions = Tools.merge({}, Configuration.options.SCATTER);
+	// determines how sensitive the grid is for hover-selection
 	gridlineThreshold: Number;
 
 	constructor(holder: Element, configs: ChartConfig<ScatterChartOptions>) {
@@ -48,9 +49,13 @@ export class ScatterChart extends BaseAxisChart {
 		// Hide the overlay
 		this.chartOverlay.hide();
 
-
 		// Dispatch the load event
 		this.dispatchEvent("load");
+
+		// check if gridline tooltips are set on
+		if (this.options.tooltip.gridline) {
+			this.addGridXEventListener();
+		}
 	}
 
 	getLegendType() {
@@ -244,6 +249,24 @@ export class ScatterChart extends BaseAxisChart {
 	}
 
 	/**
+	 * Sets the threshold for the gridline tooltips. On resize, the threshold needs to be
+	 * updated.
+	 */
+	setGridlineThreshold() {
+		const allTicks = this.svg.selectAll(".x.grid");
+
+		// select the first and the second tick to calculate the distance between
+		const first = allTicks.select(".tick");
+		const second = allTicks.select(".tick + g.tick");
+
+		// get space between axis grid ticks
+		const gridSpacing = (+Tools.getTranslationValues(second.node()).tx - +Tools.getTranslationValues(first.node()).tx);
+
+		// adjust the threshold for the tooltips
+		this.gridlineThreshold = gridSpacing * Configuration.tooltip.axisTooltip.axisThreshold;
+	}
+
+	/**
 	 * Retrieves the d3 selection of all points with domain label
 	 * @param domain The X-Axis domain label associated with data group
 	 */
@@ -260,4 +283,54 @@ export class ScatterChart extends BaseAxisChart {
 		return this.svg.selectAll("circle.dot")
 		.filter(function() { return Number(this.attributes.cx.value) === xPosition; });
 	}
+
+	/**
+	 * Adds the listener on the X grid to trigger multiple point tooltips along the x axis.
+	 */
+	addGridXEventListener() {
+		const self = this;
+		const grid = Tools.appendOrSelect(this.svg, "rect.chart-grid-backdrop");
+
+		this.setGridlineThreshold();
+
+		grid
+		.on("mousemove", function() {
+			const chartContainer = this.parentNode;
+			const pos = mouse(chartContainer);
+
+			const allgridlines = self.svg.selectAll(".x.grid .tick");
+			// remove the styling on the lines
+			allgridlines.classed("active", false);
+
+			const activeGridlines = self.getActiveGridLines(pos);
+			if (!activeGridlines) {
+				self.hideTooltip();
+				return;
+			}
+
+			// set active class to control dasharray and theme colors
+			activeGridlines
+			.classed("active", true);
+
+			// get the items that should be highlighted
+			let highlightItems;
+			activeGridlines.each(function(d) {
+				if (d) {
+					// prioritize using domain to get all points
+					// in case there are axis lines without labels
+					highlightItems = self.getDataWithDomain(d);
+				} else {
+					const translatePos = Tools.getTranslationValues(this);
+					highlightItems = self.getDataWithXValue(+translatePos.tx - 0.5);
+				}
+			});
+			self.showTooltip(highlightItems.data());
+		})
+		.on("mouseout", function() {
+			self.svg.selectAll(".x.grid .tick")
+			.classed("active", false);
+			self.hideTooltip();
+		});
+	}
+
 }
